Update the unit test data sets
The previously used data for the RDF-based package creation unit test is
removed and new data is provided.  The test will be updated in a
forthcoming submission to use the new data.  Data is also added for the
folder-plus-rules-based unit test, which will also be included in the
forthcoming submission.